moved  `reshape_embedding` function to embedding.py and added some docstrings to rewrites.py
